refactor: delete unused comments
