+++ src/Components/Documentation/FieldsList.tsx
@@ -9,9 +9,7 @@ import {
   activePoint,
   exampleText,
   flexColumnCenter,
-  flexRowCenter,
   hideBlock,
-  returnTitle,
   schemaHeading,
   sectionHeading,
   sectionSubHeading,
@@ -21,7 +19,6 @@ import {
 } from './styles';
 import { Field, Type } from '../../common-types/schema.types';
 import KeyboardArrowRightIcon from '@mui/icons-material/KeyboardArrowRight';
-import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
 import { useState } from 'react';
 import {
   DEFAULT_CURRENT_FIELD,
@@ -35,8 +32,6 @@ import UIContent from '../../assets/UIStrings.json';
 import { getFieldTypeName } from '../../utils/getFieldTypeName';
 
 const FieldsList: React.FC<FieldsListProps> = ({
-  // prevFiled,
-  // prevFiledType,
   currentFiledType,
   currentFiled,
   types,
@@ -56,44 +51,27 @@ const FieldsList: React.FC<FieldsListProps> = ({
     setSectionIsOpen(false);
   };
 
-  const handleBackToQueries = () => {
-    console.log('ok');
-    // setSectionIsOpen(false);
-    // dispatch(setDocsIsOpen(false));
-    // dispatch(setDocsIsOpen(true));
-  };
-
   const theme = useTheme();
   const isMobileView = useMediaQuery(theme.breakpoints.down('sm'));
 
   return (
-    <Box sx={wrapperDocsSection}>
-      {currentFiledType && (
-        <Box
-          sx={
-            isMobileView
-              ? Object.assign(
-                  {},
-                  flexColumnCenter,
-                  sectionIsOpen ? showBlock : hideBlock
-                )
-              : flexColumnCenter
-          }
-        >
-          <Box sx={wrapperNextDocsSection}>
+    <>
+      <Box sx={wrapperDocsSection}>
+        {currentFiledType && (
+          <Box
+            sx={
+              isMobileView
+                ? Object.assign(
+                    {},
+                    wrapperNextDocsSection,
+                    sectionIsOpen ? showBlock : hideBlock
+                  )
+                : wrapperNextDocsSection
+            }
+          >
             <Typography sx={sectionHeading} variant="h4">
               {UIContent[DOCS_HEADERS.Type_details][language]}
             </Typography>
-
-            <Box sx={flexRowCenter} onClick={() => handleBackToQueries()}>
-              <IconButton>
-                <ArrowBackIosIcon />
-              </IconButton>
-              <Typography sx={returnTitle} variant="h4">
-                {/* {`Back to ${prevFiledType}`} */}
-                {`Back to QUERIES`}
-              </Typography>
-            </Box>
 
             <Typography sx={sectionSubHeading}>
               {currentFiled.description}
@@ -146,35 +124,33 @@ const FieldsList: React.FC<FieldsListProps> = ({
             })}
             <Typography variant="subtitle1" sx={exampleText}>{`}`}</Typography>
           </Box>
-        </Box>
-      )}
+        )}
 
-      {currentNextFiledType && (
-        <Box
-          sx={
-            isMobileView
-              ? Object.assign(
-                  {},
-                  wrapperDocsSection,
-                  sectionIsOpen ? hideBlock : showBlock
-                )
-              : wrapperDocsSection
-          }
-        >
-          <Box sx={flexColumnCenter}>
-            <FieldsList
-              currentFiledType={currentNextFiledType}
-              currentFiled={currentNextFiled}
-              types={types}
-              prevFiled={currentFiled}
-              prevFiledType={currentFiledType}
-            />
+        {currentNextFiledType && (
+          <Box
+            sx={
+              isMobileView
+                ? Object.assign(
+                    {},
+                    wrapperDocsSection,
+                    sectionIsOpen ? hideBlock : showBlock
+                  )
+                : wrapperDocsSection
+            }
+          >
+            <Box sx={flexColumnCenter}>
+              <FieldsList
+                currentFiledType={currentNextFiledType}
+                currentFiled={currentNextFiled}
+                types={types}
+              />
 
-            <ArgsList currentFiled={currentNextFiled} />
+              <ArgsList currentFiled={currentNextFiled} />
+            </Box>
           </Box>
-        </Box>
-      )}
-    </Box>
+        )}
+      </Box>
+    </>
   );
 };
